more work

+++ resources/js/lycheeOrg/photoView.js
@@ -2,7 +2,7 @@
 
 export default { photoView };
 
-export function photoView(detailsOpen_val, isFullscreen_val, has_description_val, overlayType_val) {
+export function photoView(detailsOpen_val, isFullscreen_val, has_description_val, overlayType_val, canEdit_val = false) {
 	return {
 		detailsOpen: detailsOpen_val,
 		isFullscreen: isFullscreen_val,
@@ -10,6 +10,7 @@ export function photoView(detailsOpen_val, isFullscreen_val, has_description_val
 		overlayType: overlayType_val,
 		editOpen: false,
 		donwloadOpen: false,
+		canEdit: canEdit_val,
 
 		silentToggle(elem) {
 			this[elem] = !this[elem];
@@ -46,51 +47,65 @@ export function photoView(detailsOpen_val, isFullscreen_val, has_description_val
 			}
 			console.log(document.activeElement.nodeName);
 
-			// del (46) or backspace (8)
-			if (event.ctrlKey && (event.keyCode === 46 || event.keyCode === 8)) {
-				this.$wire.delete();
+			// f
+			if (event.keyCode === 70) {
+				this.silentToggle("isFullscreen");
+				return;
+			}
+
+			// o
+			if (event.keyCode === 79) {
+				this.rotateOverlay();
+				return;
 			}
 
 			// i
 			if (event.keyCode === 73) {
 				this.detailsOpen = !this.detailsOpen;
 				this.editOpen = false;
+				return;
+			}
+
+			if (!this.canEdit) {
+				console.log("can't edit.");
+				return;
+			}
+
+			// del (46) or backspace (8)
+			if (event.ctrlKey && (event.keyCode === 46 || event.keyCode === 8)) {
+				this.$wire.delete();
+				return;
 			}
 
 			// e
 			if (event.keyCode === 69) {
 				this.detailsOpen = false;
 				this.editOpen = !this.editOpen;
-			}
-
-			// f
-			if (event.keyCode === 70) {
-				this.silentToggle("isFullscreen");
+				return;
 			}
 
 			// m
 			if (event.keyCode === 77) {
 				this.$wire.move();
-			}
-
-			// o
-			if (event.keyCode === 79) {
-				this.rotateOverlay();
+				return;
 			}
 
 			// s
 			if (event.keyCode === 83) {
 				this.$wire.set_star();
+				return;
 			}
 
-			// left arrow
+			// ctrl + left arrow
 			if (event.keyCode === 37 && event.ctrlKey) {
 				this.$wire.rotate_ccw();
+				return;
 			}
 
-			// right arrow
+			// ctrl + right arrow
 			if (event.keyCode === 39 && event.ctrlKey) {
 				this.$wire.rotate_cw();
+				return;
 			}
 		},
 	};
